requirements.txt
fastapi
uvicorn
streamlit

groq

requests
python-dotenv

numpy
scikit-learn
pandas

sentence-transformers
transformers
torch

faiss-cpu

arxiv

networkx
matplotlib
bertopic
umap-learn
hdbscan

boto3